requirements.txt
google-cloud-bigquery==3.21.0
aiohttp==3.9.5
pandas==2.2.2
protobuf>=3.19.5
db-dtypes
